# sandaga-backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sandaga App
  description: >
    'This is a ***Sandaga app API***.'
  termsOfService: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  contact:
    email: info@sandaga.com
  license:
    name: MIT
    url: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://sandaga.swagger.io/v2
...
paths:
  /api/v1/users:
    post:
      tags:
        - user
      summary: Signup the a new customer (user)
      description: Creates a new customer (user), who can login and do the shopping.
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedInUser'
  /api/v1/auth/token:
    post:
      tags:
        - user
      summary: Signin the customer (user)
      description: Signin the customer (user) that generates the JWT (access token) and refresh token, which can be used for accessing APIs.
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInReq'
      responses:
        200:
          description: For user sign-in. Once successful, user receives the access and refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedInUser'
    delete:
      tags:
        - user
      summary: Signouts the customer (user)
      description: Signouts the customer (user). It removes the refresh token from DB. Last issued JWT should be removed from client end that if not removed last for given expiration time.
      operationId: signOut
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshToken'
      responses:
        202:
          description: Accepts the request for logout.
  /api/v1/auth/token/refresh:
    post:
      tags:
        - user
      summary: Provides new JWT based on valid refresh token.
      description: Provides new JWT based on valid refresh token.
      operationId: getAccessToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshToken'
      responses:
        200:
          description: For successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedInUser'
  /api/v1/auth/token/verify/{code}:
    get:
      tags:
        - user
      summary: Token for verification new register user.
      description: Token for verification new register user.
      operationId: verifyUserByToken
      parameters:
        - name: code
          in: path
          description: Token for verification
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyEntity'
  /api/v1/auth/forgot-password:
    post:
      tags:
        - user
      summary: Process email for user password forgot.
      description: Process email for user password forgot.
      operationId: forgotUserPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordUser'
      responses:
        200:
          description: For successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyEntity'
  /api/v1/auth/reset-password:
    post:
      tags:
        - user
      summary: Reset user password.
      description: Reset user password.
      operationId: resetUserPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordUser'
      responses:
        200:
          description: For successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyEntity'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: string
        addressId:
          type: string
        accountType:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
      xml:
        name: User
...
    SignInReq:
      description: Request body for Sign-in
      type: object
      properties:
        email:
          description: email of the User
          type: string
        password:
          description: password of the User
          type: string
    SignedInUser:
      description: Signed-in user information
      type: object
      properties:
        refreshToken:
          description: Refresh Token
          type: string
        accessToken:
          description: JWT Token aka access token
          type: string
        firstname:
          description: User firstname
          type: string
        lastname:
          description: User lastname
          type: string
        email:
          description: User email
          type: string
        role:
          description: User role
          type: string
        userId:
          description: User Identifier
          type: string
        accountType:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
    RefreshToken:
      description: Contains the refresh token
      type: object
      properties:
        refreshToken:
          description: Refresh Token
          type: string
    ForgotPasswordUser:
      description: Forgot password user information
      type: object
      properties:
        email:
          description: User email
          type: string
    ResetPasswordUser:
      description: Reset password user information
      type: object
      properties:
        token:
          description: Reset password token
          type: string
        password:
          description: User new password
          type: string
...
    EmptyEntity:
      type: object
      xml:
        name: EmptyEntity